chore: Add Dependabot configuration file
This commit adds a Dependabot configuration file to the repository. The file specifies the package ecosystems to update and the location of the package manifests. The configuration is set to update npm packages on a weekly basis. This addition will enable automated dependency updates and help keep the project up to date with the latest package versions.